create_results updated to check if posts already exist in db before writing new posts

+++ db/migrate/20200325215210_create_posts.rb
@@ -5,7 +5,7 @@ class CreatePosts < ActiveRecord::Migration[6.0]
       t.integer :author_id
       t.integer :origin_id
       t.integer :likes
-      t.decimal :popularity
+      t.float :popularity
       t.integer :reads
       t.text :tags, array: true
       t.timestamps
